+++ server/tests/it/main.rs
@@ -27,3 +27,5 @@ mod ws_errors;
 mod ws_get;
 mod ws_get_unauthorized_latency;
 mod ws_unsub;
+
+mod import_oad;